refactor: address a few code duplications
