+++ src/utils/index.ts
@@ -3,7 +3,7 @@ export { clearFolder } from './clearFolder';
 export { jsonSchemaToTsConst } from './jsonSchemaToTsConst';
 export { convertOpenApiParameters } from './convertOpenApiParameters';
 export { convertOpenApiToJsonSchema } from './convertOpenApiToJsonSchema';
-export { makeJsonSchemaFile } from './makeJsonSchemaFile';
+export { makeJsonSchemaFiles } from './makeJsonSchemaFiles';
 export { refToPath } from './refToPath';
 export {
   REF_SYMBOL,
@@ -16,7 +16,7 @@ export type {
   JSONSchema,
   OpenApiSchema,
   SchemaPatcher,
-  InlinedRefs,
   SchemaMetaInfo,
   SchemaMetaInfoMap,
 } from './types';
+export { addSchemaToGenerationMap } from './addSchemaToGenerationMap';
