+++ src/utils/types.ts
@@ -2,7 +2,6 @@ import type { JSONSchema4, JSONSchema6, JSONSchema7 } from 'json-schema';
 export type JSONSchema = JSONSchema4 | JSONSchema6 | JSONSchema7;
 export type OpenApiSchema = Record<string, any>;
 export type SchemaPatcher = (params: { schema: JSONSchema }) => void;
-export type InlinedRefs = Map<string, JSONSchema>;
 export type SchemaMetaInfo = {
   schemaName: string;
   schemaAbsoluteDirName: string;
